fillPage will use more safety method to submit data
